backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.11
django-cors-headers==4.3.0
python-decouple==3.8
Pillow==12.1.1
celery==5.3.4
redis==5.0.1
django-filter==23.3
drf-nested-routers==0.93.4
requests==2.31.0
reportlab==4.2.5


